// src/app/models/puntos-de-interes.ts
// Generated by https://quicktype.io

export interface PuntoDeInteres {
  idPuntoInteres:       number;
  latitud:              number;
  longitud:             number;
  telefono:             string;
  tipo:                 string;
  descripcion:          string;
  idProveedor:          number;
  usuarioCreacion:      number;
  fechaCreacion:        string;
  fechaCreacionUtc:     string;
  usuarioModificacion:  null;
  fechaModificacion:    null;
  fechaModificacionUtc: null;
  proveedor:            null;
  idProveedorWeb:       number;
  idUsuarioWeb:         number;
  proveedorWeb:         null;
}